refactor(server): clean-up codes
