+++ server/internal/utils/logger/logger.go
@@ -12,10 +12,24 @@ type Logger struct {
 	logr.Logger
 }
 
+// LogLevel type
+type LogLevel int8
+
+// Logging levels
+const (
+	DebugLevel LogLevel = iota - 1
+	InfoLevel
+	WarnLevel
+	ErrorLevel
+	DPanicLevel
+	PanicLevel
+	FatalLevel
+)
+
 // Options is a struct for configuring the Logger.
 type Options struct {
 	Development bool
-	Level       int
+	Level       LogLevel
 	Debug       bool
 }
 
@@ -24,25 +38,38 @@ func New(opts Options) *Logger {
 	var zapLogger *zap.Logger
 	var err error
 
-	if opts.Development {
-		zapLogger, err = zap.NewDevelopment()
-		if err != nil {
-			panic(err)
-		}
-	} else {
-		zapLogger, err = zap.NewProduction()
-		if err != nil {
-			panic(err)
-		}
+	config := zap.Config{
+		Level:       zap.NewAtomicLevelAt(zapcore.Level(opts.Level)),
+		Development: opts.Development,
+		Encoding:    "json",
+		EncoderConfig: zapcore.EncoderConfig{
+			TimeKey:        "timestamp",
+			LevelKey:       "level",
+			NameKey:        "logger",
+			CallerKey:      "caller",
+			MessageKey:     "message",
+			StacktraceKey:  "stacktrace",
+			LineEnding:     zapcore.DefaultLineEnding,
+			EncodeLevel:    zapcore.LowercaseLevelEncoder,
+			EncodeTime:     zapcore.ISO8601TimeEncoder,
+			EncodeDuration: zapcore.StringDurationEncoder,
+			EncodeCaller:   zapcore.ShortCallerEncoder,
+		},
+		OutputPaths:      []string{"stdout"},
+		ErrorOutputPaths: []string{"stderr"},
+		InitialFields:    map[string]interface{}{"platform": "server"},
 	}
 
 	if opts.Debug {
-		zapLogger = zapLogger.WithOptions(zap.AddStacktrace(zap.ErrorLevel))
+		config.Encoding = "console"
+		config.EncoderConfig.EncodeLevel = zapcore.CapitalLevelEncoder
+		config.EncoderConfig.EncodeTime = zapcore.TimeEncoderOfLayout("2006-01-02 15:04:05.000")
 	}
 
-	zapLogger = zapLogger.WithOptions(zap.IncreaseLevel(zap.LevelEnablerFunc(func(lvl zapcore.Level) bool {
-		return lvl >= zapcore.Level(opts.Level)
-	})))
+	zapLogger, err = config.Build()
+	if err != nil {
+		panic(err)
+	}
 
 	logger := zapr.NewLogger(zapLogger)
 
@@ -51,3 +78,6 @@ func New(opts Options) *Logger {
 	}
 }
 
+func (l *Logger) Warn(msg string, keysAndValues ...interface{}) {
+	l.Logger.V(0).Info(msg, keysAndValues...)
+}
