+++ server/main.go
@@ -4,6 +4,7 @@ import (
 	"Deployer/internal/utils/config"
 	"Deployer/internal/utils/logger"
 	"Deployer/internal/web"
+	"github.com/labstack/echo/v4/middleware"
 )
 
 func main() {
@@ -13,9 +14,13 @@ func main() {
 
 	log := logger.New(logger.Options{
 		Development: true,
-		Level:       0,
+		Level:       logger.InfoLevel,
 	})
 
 	server := web.NewServer(webServerConfig, log)
+
+	// Add middleware for gracefully handling panics
+	server.Use(middleware.Recover())
+
 	server.Start()
 }
